+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.4 - 2021-06-08
+
+* Ensure `createHitmap` returns a sorted hitmap. This fixes a potential issue with
+  ignore line annotations.
+
 ## 1.0.3 - 2021-05-25
 
 * Updated dependency on `vm_service` package from `^6.1.0` to `>=6.1.0 <8.0.0`.